programming really is crazy huh - forcing integers before strings so the whole demand sector doesn't break

+++ PREPARE-TIMES-NZ/src/prepare_times_nz/stage_4/demand_projections/demand_drivers.py
@@ -324,7 +324,7 @@ def make_demand_drivers(df, scenario):
     df = df.drop("Scenario", axis=1)
 
     # pivot out. First do this weird thing to the year vars
-    df["Year"] = r"\~" + df["Year"].astype(str)
+    df["Year"] = r"\~" + df["Year"].astype(int).astype(str)
     # pivot
     df = df.pivot(
         index=["Region", "Driver"], columns="Year", values="Index"
